Based on the Ray, the color changed a little bit to make a diff between y-Touch and x-Touch

+++ Player.h
@@ -20,8 +20,8 @@ typedef struct {
 Player* createPlayer(int posX, int posY, int size, int FOV);
 void showPlayer(Player* player, SDL_Renderer* renderer);
 void movePlayer(Player* player, int speed, const Uint8* keyState, SDL_Rect* playerRect, SDL_FPoint* viewPoint, float angle);
-void calculateAnglePointPlayer(const Player *firstPlayer, float angle, float x, float y, float xPlayerPosView,
+void calculateAnglePointPlayer(const Player *firstPlayer, int angle, float x, float y, float xPlayerPosView,
                                float yPlayerPosView, SDL_FPoint *viewPointWall, SDL_FPoint *viewPointPlayer);
-void changeAngle(float* angle, const Uint8* keyState);
+void changeAngle(int* angle, const Uint8* keyState);
 
 #endif //RAYCAST3DPROJECT_PLAYER_H
